Created a new plugin action `create-object-scoped` that is identical to the `create-object` plugin action, but with a bug fix.
The current `create-object` plugin action resolves the object data without passing the variable scope of the execution. This means that if we try to resolve process variables, it will look at process variables of ALL the processes that belong to a case. This will cause issues if your process calls a sub-process multiple times. Process variables that are set inside of these sub-processes will have the same name, causing conflicts/errors.

The new `create-object-scoped` plugin action passes the variable scope correctly, so that the process variables in these sub processes are isolated.

Unfortunately we can't change the original `create-object` because that would be a breaking change.